Add i2gw/provider.go file for provider related types
